requirements.txt
Django
django-redis
django-environ
requests